Tests suppression impossible si ID inexistant validé

+++ erp_back/src/main/java/fr/iut_unilim/erp_back/controllers/SaeController.java
@@ -48,6 +48,9 @@ public class SaeController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteSae(@PathVariable Long id) {
+        if (!saeRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         saeService.deleteSaeById(id);
         return ResponseEntity.ok().body("La SAE a bien été supprimée");
     }
